move AboutUsActivity to aboutUs package
