+++ app/src/main/java/com/example/android/psiak/ui/main/MainActivity.java
@@ -21,6 +21,7 @@ import android.widget.Toast;
 import com.example.android.psiak.R;
 import com.example.android.psiak.data.network.FirebaseRepository;
 import com.example.android.psiak.model.DogFirebase;
+import com.example.android.psiak.ui.aboutUs.AboutUsActivity;
 import com.example.android.psiak.ui.addAnimal.AddAnimalActivity;
 import com.mindorks.placeholderview.SwipeDecor;
 import com.mindorks.placeholderview.SwipePlaceHolderView;
