+++ app/src/main/res/layout/activity_about_us.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.android.psiak.ui.main.AboutUsActivity">
+    tools:context="com.example.android.psiak.ui.aboutUs.AboutUsActivity">
 
 
     <android.support.design.widget.AppBarLayout
